+++ php/WpDateTimeInterface.php
@@ -5,9 +5,13 @@ namespace Rarst\WordPress\DateTime;
 /**
  * @see WpDateTimeTrait
  */
-interface WpDateTimeInterface {
+interface WpDateTimeInterface extends \DateTimeInterface {
 
 	public static function createFromPost( $post, $field = 'date' );
 
 	public function formatI18n( $format );
+
+	public function formatDate();
+
+	public function formatTime();
 }